+++ Sources/SwiftGen/main.swift
@@ -13,6 +13,7 @@ import SwiftGenKit
 
 // MARK: - Main
 
+// swiftlint:disable:next closure_body_length
 let main = Group {
   $0.noCommand = { path, group, parser in
     if parser.hasOption("help") {
@@ -23,17 +24,24 @@ let main = Group {
     }
   }
   $0.group("config", "manage and run configuration files") {
-    $0.addCommand("lint", "Lint the configuration file", ConfigCLI.lint)
-    $0.addCommand("run", "Run commands listed in the configuration file", ConfigCLI.run)
-    $0.addCommand("init", "Create an initial configuration file", ConfigCLI.create)
-    $0.addCommand("doc", "Open the documentation for the configuration file on GitHub", ConfigCLI.doc)
+    $0.addCommand("lint", "lint the configuration file", ConfigCLI.lint)
+    $0.addCommand("run", "run commands listed in the configuration file", ConfigCLI.run)
+    $0.addCommand("init", "create an initial configuration file", ConfigCLI.create)
+    $0.addCommand("doc", "open the documentation for the configuration file on GitHub", ConfigCLI.doc)
   }
 
-  $0.group("templates", "manage custom templates") {
-    $0.addCommand("list", "list bundled and custom templates", TemplatesCLI.list)
-    $0.addCommand("which", "print path of a given named template", TemplatesCLI.which)
-    $0.addCommand("cat", "print content of a given named template", TemplatesCLI.cat)
+  $0.group("template", "manage custom templates") {
+    $0.addCommand("list", "list bundled and custom templates", TemplateCLI.list)
+    $0.addCommand("which", "print path of a given named template", TemplateCLI.which)
+    $0.addCommand("cat", "print content of a given named template", TemplateCLI.cat)
   }
+  // Deprecated: Remove this in SwiftGen 7.0
+  $0.group("templates", "DEPRECATED - use `template` subcommand instead") {
+    $0.addDeprecatedCommand("list", replacement: "template list", TemplateCLI.list)
+    $0.addDeprecatedCommand("which", replacement: "template which", TemplateCLI.which)
+    $0.addDeprecatedCommand("cat", replacement: "template cat", TemplateCLI.cat)
+  }
+  // Deprecation end
 
   for cmd in ParserCLI.allCommands {
     $0.addCommand(cmd.name, cmd.description, cmd.command())
@@ -48,3 +56,21 @@ main.run(
   SwiftGenKit v\(Version.swiftGenKit))
   """
 )
+
+// Deprecated: Remove this in SwiftGen 7.0
+extension Group {
+  struct DeprecatedCommand: CommandType {
+    let wrappedCommand: CommandType
+    let replacement: String
+
+    func run(_ parser: ArgumentParser) throws {
+      logMessage(.warning, "This command is deprecated in favor of `\(replacement)`")
+      try wrappedCommand.run(parser)
+    }
+  }
+  public func addDeprecatedCommand(_ name: String, replacement: String, _ command: CommandType) {
+    let depCmd = DeprecatedCommand(wrappedCommand: command, replacement: replacement)
+    addCommand(name, "DEPRECATED - use `\(replacement)` instead", depCmd)
+  }
+}
+// Deprecation end